# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is the backend API for the Tribute application, which integrates with Telegram for user interaction and authentication. It uses Telegram's `initData` for secure, stateless authentication.",
        "title": "Tribute Backend API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/user/Tribute-back/issues",
            "email": "support@tribute.app"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/add-bot": {
            "post": {
                "security": [
                    {
                        "TgAuth": []
                    }
                ],
                "description": "Adds a new Telegram channel for the authenticated user. This allows the system to associate a bot with the user's account.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tribute"
                ],
                "summary": "Add a new Bot/Channel",
                "parameters": [
                    {
                        "description": "The username of the bot/channel to add.",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AddBotRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created - The bot was added successfully.",
                        "schema": {
                            "$ref": "#/definitions/dto.AddBotResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request - The request body is invalid.",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - The Authorization header is missing or invalid.",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden - The provided initData is invalid or expired.",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error - e.g., bot with this username already exists.",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/check-verified-passport": {
            "post": {
                "description": "**PUBLIC ENDPOINT.** This endpoint is intended to be called by Telegram in response to an admin clicking a button in the verification chat. It should not be called directly by the frontend. It processes verification approvals and rejections.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Webhooks"
                ],
                "summary": "Telegram Verification Webhook",
                "parameters": [
                    {
                        "description": "The callback query update sent by Telegram.",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.TelegramUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success - The callback was processed.",
                        "schema": {
                            "$ref": "#/definitions/dto.StatusResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request - The payload from Telegram is malformed.",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error - Failed to process the callback data.",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/create-subscribe": {
            "post": {
                "security": [
                    {
                        "TgAuth": []
                    }
                ],
                "description": "Allows an authenticated user (the subscriber) to pay for and subscribe to another user (the creator). This action creates a payment record and updates the creator's earnings.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tribute"
                ],
                "summary": "Subscribe to an Author",
                "parameters": [
                    {
                        "description": "The ID of the user to subscribe to and the price.",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateSubscribeRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created - The subscription was successful.",
                        "schema": {
                            "$ref": "#/definitions/dto.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request - The request body is invalid.",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - The Authorization header is missing or invalid.",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden - The provided initData is invalid or expired.",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error - e.g., the creator has no subscription tier, or the price is incorrect.",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/dashboard": {
            "get": {
                "security": [
                    {
                        "TgAuth": []
                    }
                ],
                "description": "Retrieves all data for the main dashboard screen. The user is identified via the `initData` in the Authorization header. If the user does not exist in the database, empty data is returned, which is a normal flow indicating the user needs to complete onboarding via the `/onboard` endpoint.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tribute"
                ],
                "summary": "Get Dashboard Data",
                "responses": {
                    "200": {
                        "description": "Successfully retrieved dashboard data (empty if user not onboarded).",
                        "schema": {
                            "$ref": "#/definitions/dto.DashboardResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - The Authorization header is missing or invalid.",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden - The provided initData is invalid or expired.",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error - An unexpected error occurred.",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/onboard": {
            "put": {
                "security": [
                    {
                        "TgAuth": []
                    }
                ],
                "description": "Creates a user record if one doesn't exist, or updates an existing user to mark them as onboarded. This is the first endpoint a new user should call. It is idempotent.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tribute"
                ],
                "summary": "Onboard a User",
                "responses": {
                    "200": {
                        "description": "Success - The user already existed and has been marked as onboarded.",
                        "schema": {
                            "$ref": "#/definitions/dto.OnboardResponse"
                        }
                    },
                    "201": {
                        "description": "Created - A new user was created and marked as onboarded.",
                        "schema": {
                            "$ref": "#/definitions/dto.OnboardResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - The Authorization header is missing or invalid.",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden - The provided initData is invalid or expired.",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error - An unexpected error occurred.",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/publish-subscription": {
            "put": {
                "security": [
                    {
                        "TgAuth": []
                    }
                ],
                "description": "Allows an author to create or update their public subscription details (title, description, price). This is an idempotent operation. The user must have at least one channel added via `/add-bot` to use this.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tribute"
                ],
                "summary": "Publish or Update a Subscription Tier",
                "parameters": [
                    {
                        "description": "The details of the subscription tier to publish.",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.PublishSubscriptionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success - The subscription was published or updated successfully.",
                        "schema": {
                            "$ref": "#/definitions/dto.PublishSubscriptionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request - The request body is invalid.",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - The Authorization header is missing or invalid.",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden - The provided initData is invalid or expired.",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error - e.g., the user has no channels.",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.AddBotRequest": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "bot-username": {
                    "type": "string"
                }
            }
        },
        "dto.AddBotResponse": {
            "type": "object",
            "properties": {
                "channel": {
                    "$ref": "#/definitions/dto.ChannelDTO"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "dto.StatusResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                }
            }
        },
...
        "dto.TelegramUpdate": {
            "type": "object",
            "properties": {
                "callback_query": {
                    "$ref": "#/definitions/dto.CallbackQuery"
                },
                "update_id": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "TgAuth": {
            "description": "**Authentication for Telegram Web Apps.** \\\nEnter your token in the format: `TgAuth \u003cinitData\u003e`. \\\nThe `\u003cinitData\u003e` string is provided by the Telegram client when the web app is opened.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}